chore - rename interactiveEditor to inlineChat (#184422)
