+++ src/vs/workbench/contrib/inlineChat/browser/inlineChat.css
