+++ src/vs/workbench/contrib/inlineChat/browser/utils.ts
